refactor: rewrite some stuff for better organizing and cleaner code
This commit simplifies the dependencies to make the script more
lightweight and it also introduces new options - date bounds. You can
set min-max times of day where the bot is allowed to work, as well as
days of the week. There's also a field for mentioning someone when the
endpoint goes down. These are all optional.

The script now knows when the error was local or not